Refactor AI model fallback logic into a new utility module for improved maintainability and extend model options in aiController.
